requirements.txt
numpy==1.22.1
pandas==1.4.0
pandas-gbq==0.19.2
google-cloud-storage==2.12.0
gcsfs==2023.9.2
pytest==7.2.0